test: Move key parsing/dedup tests to new file
